+++ src/price/main/service/CurrencyService.java
@@ -1,0 +1,14 @@
+package price.main.service;
+
+import price.main.enums.Digit;
+
+/**
+ * may be use for different currencies
+ */
+public interface CurrencyService {
+    String getSingleCurrencyNominative(Digit digit);
+
+    String getSingleCurrencyDative(Digit digit);
+
+    String getMultipleCurrency(Digit digit);
+}
